Introduce Process interface and ProcessBase for parallel workers
Splits the parallel worker into a process-creation-agnostic half
(ProcessBase: the TCP/NDJSON connection and timeout timer) and the
process-creation half. The Process interface is what ParallelAnalyser
and ProcessPool talk to; the existing react/child-process implementation
is renamed to SpawnedProcess. Sole implementation for now — a
pcntl_fork() based one follows.
