+++ src/Parallel/Process.php
@@ -2,151 +2,36 @@
 
 namespace PHPStan\Parallel;
 
-use PHPStan\ShouldNotHappenException;
-use React\EventLoop\LoopInterface;
-use React\EventLoop\TimerInterface;
 use React\Stream\ReadableStreamInterface;
 use React\Stream\WritableStreamInterface;
 use Throwable;
-use function fclose;
-use function rewind;
-use function sprintf;
-use function stream_get_contents;
-use function tmpfile;
 
-final class Process
+/**
+ * A parallel analysis worker as seen by ParallelAnalyser / ProcessPool.
+ *
+ * Implementations differ only in how the worker process comes to life:
+ * SpawnedProcess spawns a fresh PHP process via react/child-process,
+ * ForkedProcess forks the already-booted main process via pcntl_fork(). Both
+ * then speak the same TCP + NDJSON protocol, so request()/quit()/
+ * bindConnection() behave identically and live in ProcessBase.
+ */
+interface Process
 {
-
-	private \React\ChildProcess\Process $process;
-
-	private ?WritableStreamInterface $in = null;
-
-	/** @var resource */
-	private $stdOut;
-
-	/** @var resource */
-	private $stdErr;
-
-	/** @var callable(mixed[] $json) : void */
-	private $onData;
-
-	/** @var callable(Throwable $exception): void */
-	private $onError;
-
-	private ?TimerInterface $timer = null;
-
-	public function __construct(
-		private string $command,
-		private LoopInterface $loop,
-		private float $timeoutSeconds,
-	)
-	{
-	}
 
 	/**
 	 * @param callable(mixed[] $json) : void $onData
 	 * @param callable(Throwable $exception): void $onError
 	 * @param callable(?int $exitCode, string $output) : void $onExit
 	 */
-	public function start(callable $onData, callable $onError, callable $onExit): void
-	{
-		$tmpStdOut = tmpfile();
-		if ($tmpStdOut === false) {
-			throw new ShouldNotHappenException('Failed creating temp file for stdout.');
-		}
-		$tmpStdErr = tmpfile();
-		if ($tmpStdErr === false) {
-			throw new ShouldNotHappenException('Failed creating temp file for stderr.');
-		}
-		$this->stdOut = $tmpStdOut;
-		$this->stdErr = $tmpStdErr;
-		$this->process = new \React\ChildProcess\Process($this->command, fds: [
-			1 => $this->stdOut,
-			2 => $this->stdErr,
-		]);
-		$this->process->start($this->loop);
-		$this->onData = $onData;
-		$this->onError = $onError;
-		$this->process->on('exit', function ($exitCode) use ($onExit): void {
-			$this->cancelTimer();
-
-			$output = '';
-			rewind($this->stdOut);
-			$output .= stream_get_contents($this->stdOut);
-
-			rewind($this->stdErr);
-			$output .= stream_get_contents($this->stdErr);
-
-			$onExit($exitCode, $output);
-			fclose($this->stdOut);
-			fclose($this->stdErr);
-		});
-	}
-
-	private function cancelTimer(): void
-	{
-		if ($this->timer === null) {
-			return;
-		}
-
-		$this->loop->cancelTimer($this->timer);
-		$this->timer = null;
-	}
+	public function start(callable $onData, callable $onError, callable $onExit): void;
 
 	/**
 	 * @param mixed[] $data
 	 */
-	public function request(array $data): void
-	{
-		$this->cancelTimer();
-		if ($this->in === null) {
-			throw new ShouldNotHappenException();
-		}
-		$this->in->write($data);
-		$this->timer = $this->loop->addTimer($this->timeoutSeconds, function (): void {
-			$onError = $this->onError;
-			$onError(new ProcessTimedOutException(sprintf('Child process timed out after %.1f seconds. Try making it longer with parallel.processTimeout setting.', $this->timeoutSeconds)));
-		});
-	}
+	public function request(array $data): void;
 
-	public function quit(): void
-	{
-		$this->cancelTimer();
-		if (!$this->process->isRunning()) {
-			return;
-		}
+	public function quit(): void;
 
-		foreach ($this->process->pipes as $pipe) {
-			$pipe->close();
-		}
-
-		if ($this->in === null) {
-			return;
-		}
-
-		$this->in->end();
-	}
-
-	public function bindConnection(ReadableStreamInterface $out, WritableStreamInterface $in): void
-	{
-		$out->on('data', function (array $json): void {
-			$this->cancelTimer();
-			if ($json['action'] !== 'result') {
-				return;
-			}
-
-			$onData = $this->onData;
-			$onData($json['result']);
-		});
-		$this->in = $in;
-		$out->on('error', function (Throwable $error): void {
-			$onError = $this->onError;
-			$onError($error);
-		});
-		$in->on('error', function (Throwable $error): void {
-			$onError = $this->onError;
-			$onError($error);
-		});
-	}
+	public function bindConnection(ReadableStreamInterface $out, WritableStreamInterface $in): void;
 
 }
